pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraud-llm-finetune"
version = "0.1.0"
description = "QLoRA fine-tuning project for fraud narrative classification"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Ahmad Whafa Azka Al Azkiyai" }
]
dependencies = [
  "transformers>=4.40",
  "peft>=0.10",
  "bitsandbytes>=0.43",
  "accelerate>=0.27",
  "datasets>=2.18",
  "trl>=0.8,<0.9",
  "wandb>=0.16",
  "torch>=2.1",
  "scipy>=1.12",
  "pyyaml>=6.0",
  "gradio>=4.0",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.5"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=src --cov-report=term-missing -v"
